requirements.txt
aiofiles
argparse
asyncio
colorama
csv
dns.resolver
geoip2
ipaddress
ipwhois
json
logging
requests
socket
subprocess
whois